+++ packages/eslint-config-next/rules/graphql.js
@@ -1,0 +1,8 @@
+module.exports = {
+  overrides: [
+    {
+      files: ['*.graphql'],
+      extends: 'plugin:@graphql-eslint/operations-recommended',
+    },
+  ],
+};