+++ packages/eslint-config-cypress/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-cypress@0.6.0...@side/eslint-config-cypress@1.0.0) (2023-03-29)
+
+**Note:** Version bump only for package @side/eslint-config-cypress
+
+# [1.0.0-alpha.1](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-cypress@0.6.0...@side/eslint-config-cypress@1.0.0-alpha.1) (2023-03-29)
+
+**Note:** Version bump only for package @side/eslint-config-cypress
+
+# [0.6.0](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-cypress@0.5.0...@side/eslint-config-cypress@0.6.0) (2023-03-29)
+
+### Features
+
+- upgrade lerna ([#492](https://github.com/reside-eng/lint-config/issues/492)) ([54c907b](https://github.com/reside-eng/lint-config/commit/54c907bc7d6dee491fbac3ecb9769284554785cf)), closes [#488](https://github.com/reside-eng/lint-config/issues/488)
+
+# [0.6.0-alpha.0](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-cypress@0.5.0...@side/eslint-config-cypress@0.6.0-alpha.0) (2023-03-29)
+
+### Features
+
+- upgrade lerna ([#488](https://github.com/reside-eng/lint-config/issues/488)) ([6b10c5e](https://github.com/reside-eng/lint-config/commit/6b10c5ea54a9e68f7e3b04499a48a85b704a93cc))
+
 # [0.5.0](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-cypress@0.4.0...@side/eslint-config-cypress@0.5.0) (2022-06-16)
 
 ### Features